Clean old comments and unused script
