--- background.html
@@ -1,12 +1,0 @@
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <title></title>
-  </head>
-  <body></body>
-  <script src="keys.js"></script>
-  <script src="auth.js"></script>
-  <script src="youTubeApiConnectors.js"></script>
-  <script src="background.js"></script>
-</html>
-
