+++ package.json
@@ -1,6 +1,7 @@
 {
   "name": "ytautoplaylist",
   "version": "1.0.0",
+  "type": "module",
   "scripts": {
     "test": "node tests/test.js"
   }
