+++ popup/popup.html
@@ -4,7 +4,6 @@
     <meta charset="utf-8">
     <title>YouTube playlist connector</title>
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.4/css/bulma.min.css">
-    <link rel="stylesheet" type="text/css" href="../style.css" >
 </head>
 <body class="p-3">
     <div class="buttons">
