+++ settings/settings.html
@@ -89,6 +89,5 @@
         </form>
     </div>
 </section>
-<script src="settings.js"></script>
 </body>
 </html>
