+++ src/constants.js
@@ -1,0 +1,24 @@
+export const TITLEFILTER = {
+  'UC7Elc-kLydl-NAV4g204pDQ': ['новости |', 'военное положение |'],
+  'UCt7sv-NKh44rHAEb-qCCxvA': ['ostronews', 'iphone'],
+  'UC8zQiuT0m1TELequJ5sp5zw': ['подкаст', 'подкаст', 'спецэфир'],
+  'UC3cJiUuZlpF-pkzqvSskTpg': ['разгоны #', 'чувс', 'книжный клуб'],
+  'UCixlrqz8w-oa4UzdKyHLMaA': ['yet another podcast'],
+  'UCn9bv143ECsDMw-kJCNN7QA': ['подземелья чикен']
+};
+
+export const BROADCASTFILTER = [
+  'UC7Elc-kLydl-NAV4g204pDQ',
+  'UCt7sv-NKh44rHAEb-qCCxvA',
+  'UCUGfDbfRIx51kJGGHIFo8Rw',
+  'UCBG57608Hukev3d0d-gvLhQ',
+  'UCjQdM9q_Vd2gBN9Xy_zRDJQ',
+  'UCKRC-fNrU-XvZTrwXN4Xxcg',
+  'UC5EgIZja1sE3IF9U_CKEVew',
+  'UCBq6jERElPLXL5cEy-sA9Tw',
+  'UC6uFoHcr_EEK6DgCS-LeTNA',
+  'UCY649zJeJVhhJa-rvWThZ2g',
+  'UCUmlB9SwrBVevETZV0wrVRw',
+  'UCQ_LYRUJzBfh-mvU14xCNMw',
+  'UCTUyoZMfksbNIHfWJjwr5aQ'
+];
