--- style.css
@@ -1,53 +1,0 @@
-/* Модальная структура документа */
-/*общие настройки для всего документа*/
-html,
-body {
-  font-family: "Open Sans", sans-serif;
-  font-size: 14px;
-  margin: 0;
-  min-height: 180px;
-  padding: 0;
-  width: 380px;
-}
-/*задаём настройки для заголовков первого уровня*/
-h1 {
-  font-family: "Menlo", monospace;
-  font-size: 22px;
-  font-weight: 400;
-  margin: 0;
-  color: #2f5876;
-}
-a:link,
-a:visited {
-  color: #000000;
-  outline: 0;
-  text-decoration: none;
-}
-/*задаём ширину картинки*/
-img {
-  width: 30px; /*ширина изображений*/
-}
-.modal-header {
-  align-items: center; /*выравнивание элементов по центру*/
-  border-bottom: 0.5px solid #dadada; /*свойства нижней разделительной линии*/
-  height: 50px;
-}
-.modal-content {
-  padding: 0 22px; /*отступы сверху и снизу, сверху и слева*/
-}
-.modal-icons {
-  border-top: 0.5px solid #dadada; /*свойства верхней разделительной линии*/
-  height: 50px;
-  width: 100%;
-}
-.logo {
-  padding: 16px; /*отступы со всех сторон*/
-}
-.logo-icon {
-  vertical-align: text-bottom; /*выравнивание по нижней части текста*/
-  margin-right: 12px; /*задётся отступ элементов от изображения*/
-}
-.version {
-  color: #444;
-  font-size: 18px;
-}
