+++ tests/test.js
@@ -1,8 +1,8 @@
-const assert = require('assert');
-const api = require('../youTubeApiConnectors.js');
+import assert from 'assert';
+import { getNewVideos, __setCallApi } from '../src/youTubeApiConnectors.js';
 
 const calls = [];
-api.__setCallApi(async (path) => {
+__setCallApi(async (path) => {
   calls.push(path);
   if (path === 'playlistItems') {
     const err = new Error('API playlistItems failed');
@@ -15,7 +15,7 @@ api.__setCallApi(async (path) => {
 });
 
 (async () => {
-  const res = await api.getNewVideos('UUstub');
+  const res = await getNewVideos('UUstub');
   assert.deepStrictEqual(res, { videos: [], pages: 1 });
   assert.deepStrictEqual(calls, ['playlistItems', 'search']);
   console.log('getNewVideos falls back to search');
